+++ Elements/src/Geometry/BBox3.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Elements.Geometry
 {
@@ -77,6 +78,76 @@ namespace Elements.Geometry
         }
 
         /// <summary>
+        /// Create the BBox3 for an Element. Elements without any geometry will return invalid boxes. 
+        /// Properties of the element that are themselves elements will not be considered.
+        /// </summary>
+        /// <param name="element">The element.</param>
+        public BBox3(Element element)
+        {
+            this.Min = new Vector3(double.MaxValue, double.MaxValue, double.MaxValue);
+            this.Max = new Vector3(double.MinValue, double.MinValue, double.MinValue);
+
+            var vertices = VerticesFromElement(element);
+            foreach (var v in vertices)
+            {
+                Extend(v);
+            }
+        }
+
+        private static List<Vector3> VerticesFromElement(Element element, bool includeElementTransform = true)
+        {
+            List<Vector3> vertices = new List<Vector3>();
+            switch (element)
+            {
+                case ModelPoints modelPoints:
+                    vertices.AddRange(modelPoints.Locations.Select(v => includeElementTransform ? modelPoints.Transform.OfPoint(v) : v));
+                    break;
+                case ModelCurve modelCurve:
+                    vertices.AddRange(modelCurve.Curve.RenderVertices().Select(v => includeElementTransform ? modelCurve.Transform.OfPoint(v) : v));
+                    break;
+                case ContentElement contentElement:
+                    vertices.AddRange(contentElement.BoundingBox.Corners().Select(v => includeElementTransform ? contentElement.Transform.OfPoint(v) : v));
+                    break;
+                case ElementInstance elementInstance:
+                    // instances of content elements are transformed on top of their base definition's orientation; 
+                    // other instances are not.
+                    var baseDefVertices = VerticesFromElement(elementInstance.BaseDefinition, elementInstance.BaseDefinition is ContentElement);
+                    vertices.AddRange(baseDefVertices.Select(elementInstance.Transform.OfPoint));
+                    break;
+                case MeshElement meshElement:
+                    vertices.AddRange(meshElement.Mesh.Vertices.Select(v => meshElement.Transform.OfPoint(v.Position)));
+                    break;
+                case GeometricElement geometricElement:
+                    vertices.AddRange(VerticesFromRepresentation(geometricElement.Representation, includeElementTransform ? geometricElement.Transform : null));
+                    break;
+                case Profile profile:
+                    vertices.AddRange(profile.Perimeter.Vertices);
+                    break;
+                default:
+                    break;
+            }
+            return vertices;
+        }
+
+        private static List<Vector3> VerticesFromRepresentation(Representation representation, Transform transform = null)
+        {
+            List<Vector3> vertices = new List<Vector3>();
+            if (transform == null)
+            {
+                transform = new Transform();
+            }
+            if (representation != null && representation.SolidOperations != null && representation.SolidOperations.Count > 0)
+            {
+                foreach (var solidOp in representation.SolidOperations)
+                {
+                    var solidTransform = (solidOp.LocalTransform ?? new Transform()).Concatenated(transform);
+                    vertices.AddRange(solidOp.Solid.Vertices.Select(v => solidTransform.OfPoint(v.Value.Point)));
+                }
+            }
+            return vertices;
+        }
+
+        /// <summary>
         /// Get a translated copy of the bounding box.
         /// </summary>
         /// <param name="translation">The translation to apply.</param>
@@ -92,6 +163,27 @@ namespace Elements.Geometry
         public Vector3 Center()
         {
             return this.Max.Average(this.Min);
+        }
+
+        /// <summary>
+        /// Get all 8 corners of this bounding box. 
+        /// Ordering is CCW bottom, then CCW top, each starting from minimum (X,Y). 
+        /// For a unit cube this would be:
+        /// (0,0,0),(1,0,0),(1,1,0),(0,1,0),(0,0,1),(1,0,1),(1,1,1),(0,1,1)
+        /// </summary>
+        /// <returns>The corners of the bounding box.</returns>
+        public List<Vector3> Corners()
+        {
+            return new List<Vector3> {
+                Min,
+                new Vector3(Max.X,Min.Y,Min.Z),
+                new Vector3(Max.X,Max.Y,Min.Z),
+                new Vector3(Min.X,Max.Y,Min.Z),
+                new Vector3(Min.X,Min.Y,Max.Z),
+                new Vector3(Max.X,Min.Y,Max.Z),
+                Max,
+                new Vector3(Min.X,Max.Y,Max.Z),
+            };
         }
 
         /// <summary>