+++ Elements/src/Geometry/CsgExtensions.cs
@@ -12,7 +12,7 @@ namespace Elements.Geometry
         /// <summary>
         /// Write the csg into a mesh.
         /// </summary>
-        internal static void Tessellate(this Csg.Solid csg, ref Mesh mesh, Transform transform = null, Color color = default(Color))
+        internal static void Tessellate(this Csg.Solid csg, ref Mesh mesh)
         {
             foreach (var p in csg.Polygons)
             {
@@ -187,19 +187,16 @@ namespace Elements.Geometry
 
             foreach (var f in solid.Faces.Values)
             {
-                if (f.Inner != null)
+                if (f.Inner != null && f.Inner.Count() > 0)
                 {
                     var tess = new Tess();
                     tess.NoEmptyPolygons = true;
 
                     tess.AddContour(f.Outer.ToContourVertexArray(f));
 
-                    if (f.Inner != null)
+                    foreach (var loop in f.Inner)
                     {
-                        foreach (var loop in f.Inner)
-                        {
-                            tess.AddContour(loop.ToContourVertexArray(f));
-                        }
+                        tess.AddContour(loop.ToContourVertexArray(f));
                     }
 
                     tess.Tessellate(WindingRule.Positive, LibTessDotNet.Double.ElementType.Polygons, 3);