+++ Elements/src/Geometry/Solids/Solid.cs
@@ -52,20 +52,29 @@ namespace Elements.Geometry.Solids
         /// <param name="perimeter">The perimeter of the lamina's faces.</param>
         public static Solid CreateLamina(IList<Vector3> perimeter)
         {
+            return CreateLamina(new Polygon(perimeter));
+        }
+
+        public static Solid CreateLamina(Polygon perimeter, IList<Polygon> voids = null)
+        {
             var solid = new Solid();
-            var loop1 = new Loop();
-            var loop2 = new Loop();
-            for (var i = 0; i < perimeter.Count; i++)
+            if (voids != null && voids.Count > 0)
             {
-                var a = solid.AddVertex(perimeter[i]);
-                var b = solid.AddVertex(perimeter[i == perimeter.Count - 1 ? 0 : i + 1]);
-                var e = solid.AddEdge(a, b);
-                loop1.AddEdgeToEnd(e.Left);
-                loop2.AddEdgeToStart(e.Right);
+                solid.AddFace(perimeter, voids);
+                solid.AddFace(perimeter.Reversed(), voids.Select(h => h.Reversed()).ToArray(), true);
             }
-            solid.AddFace(loop1);
-            solid.AddFace(loop2);
+            else
+            {
+                solid.AddFace(perimeter);
+                solid.AddFace(perimeter.Reversed(), null, true);
+            }
+
             return solid;
+        }
+
+        public static Solid CreateLamina(Profile profile)
+        {
+            return CreateLamina(profile.Perimeter, profile.Voids);
         }
 
         /// <summary>
@@ -392,6 +401,16 @@ namespace Elements.Geometry.Solids
         }
 
         /// <summary>
+        /// Get the Mesh of this Solid.
+        /// </summary>
+        public Mesh ToMesh()
+        {
+            var mesh = new Mesh();
+            this.Tessellate(ref mesh);
+            return mesh;
+        }
+
+        /// <summary>
         /// Triangulate this solid.
         /// </summary>
         /// <param name="mesh">The mesh to which the solid's tessellated data will be added.</param>
@@ -452,6 +471,8 @@ namespace Elements.Geometry.Solids
                 mesh.AddMesh(faceMesh);
             }
         }
+
+
 
         /// <summary>
         /// Triangulate this solid and pack the triangulated data into buffers
@@ -686,7 +707,7 @@ namespace Elements.Geometry.Solids
         private Loop SweepPolygonBetweenPlanes(Polygon p, Transform start, Transform end, double rotation = 0.0)
         {
             // Transform the polygon to the mid plane between two transforms
-            // then project onto the end transforms. We do this so that we 
+            // then project onto the end transforms. We do this so that we
             // do not introduce shear into the transform.
             var v = (start.Origin - end.Origin).Unitized();
             var midTrans = new Transform(end.Origin.Average(start.Origin), start.YAxis.Cross(v), v);